refactor(wallet-sdk): move user repositories, service, and cache to SDK (Phase 3.2)
Adds packages/wallet-sdk/src/user/{user-repository,user-service,user-cache}.ts
with ReadUserRepository, WriteUserRepository, ReadUserDefaultAccountRepository,
UserService, UserCache, userQueryOptions, and createUserChangeHandlers — all
framework-free. ReadUserDefaultAccountRepository keeps its positional ctor: its
only consumer is the server-side lightning-address path which constructs it
per-request with the server db and env mnemonic.

Web shims at features/user/{user-repository,user-service} re-export the SDK
classes and keep the React hook wrappers for one more chunk. user-hooks.tsx
imports UserCache and createUserChangeHandlers from the SDK instead of
defining them locally; the cache-read helpers stay local until the curated
surface lands.
