+++ apps/web-wallet/app/features/user/user-repository.ts
@@ -1,388 +1,12 @@
-import type { QueryClient } from '@tanstack/react-query';
-import type { DistributedOmit } from 'type-fest';
-import type { z } from 'zod/mini';
-import { normalizeMintUrl } from '~/lib/cashu';
-import type { Currency } from '~/lib/money';
-import type { Account, RedactedAccount } from '../accounts/account';
-import {
-  type AccountRepository,
-  useAccountRepository,
-} from '../accounts/account-repository';
-import {
-  type AgicashDb,
-  type AgicashDbAccount,
-  type AgicashDbUser,
-  isCashuAccount,
-  isSparkAccount,
-} from '../agicash-db/database';
+// Transitional re-export — moved to @agicash/wallet-sdk; removed in the import-cleanup PR.
+export {
+  ReadUserDefaultAccountRepository,
+  ReadUserRepository,
+  WriteUserRepository,
+} from '@agicash/wallet-sdk';
+import { ReadUserRepository, WriteUserRepository } from '@agicash/wallet-sdk';
+import { useAccountRepository } from '../accounts/account-repository';
 import { agicashDbClient } from '../agicash-db/database.client';
-import { CashuAccountDetailsDbDataSchema } from '../agicash-db/json-models/cashu-account-details-db-data';
-import type { SparkNetwork } from '../agicash-db/json-models/spark-account-details-db-data';
-import { SparkAccountDetailsDbDataSchema } from '../agicash-db/json-models/spark-account-details-db-data';
-import {
-  getInitializedCashuWallet,
-  getMintAuthProvider,
-} from '../shared/cashu';
-import { UniqueConstraintError } from '../shared/error';
-import { getInitializedSparkWallet } from '../shared/spark';
-import type { User } from './user';
-
-export type UpdateUser = {
-  defaultBtcAccountId?: string;
-  defaultUsdAccountId?: string | null;
-  defaultCurrency?: Currency;
-  username?: string;
-  termsAcceptedAt?: string;
-  giftCardMintTermsAcceptedAt?: string;
-};
-
-type Options = {
-  abortSignal?: AbortSignal;
-};
-
-type AccountInput = {
-  isDefault?: boolean;
-} & DistributedOmit<
-  Account,
-  | 'id'
-  | 'createdAt'
-  | 'version'
-  | 'proofs'
-  | 'keysetCounters'
-  | 'wallet'
-  | 'isOnline'
-  | 'balance'
-  | 'state'
->;
-
-export class WriteUserRepository {
-  constructor(
-    private readonly db: AgicashDb,
-    private readonly accountRepository: AccountRepository,
-  ) {}
-
-  /**
-   * Updates a user in the database.
-   * @param user - The user data to update. All specified properties will be updated.
-   * @returns The updated user.
-   * @throws Error if trying to set default_currency when no default account is set for the given currency (enforced by database constraint)
-   */
-  async update(
-    userId: string,
-    data: UpdateUser,
-    options?: { abortSignal?: AbortSignal },
-  ): Promise<User> {
-    const query = this.db
-      .from('users')
-      .update({
-        default_btc_account_id: data.defaultBtcAccountId,
-        default_usd_account_id: data.defaultUsdAccountId,
-        default_currency: data.defaultCurrency,
-        username: data.username,
-        terms_accepted_at: data.termsAcceptedAt,
-        gift_card_mint_terms_accepted_at: data.giftCardMintTermsAcceptedAt,
-      })
-      .eq('id', userId)
-      .select();
-    if (options?.abortSignal) {
-      query.abortSignal(options.abortSignal);
-    }
-
-    const { data: updatedUser, error } = await query.single();
-
-    if (error) {
-      if (error.code === '23505') {
-        throw new UniqueConstraintError(error.message);
-      }
-
-      throw new Error('Failed to update user', { cause: error });
-    }
-
-    return ReadUserRepository.toUser(updatedUser);
-  }
-
-  /**
-   * Inserts a user into the database. If the user already exists, it updates the user.
-   * @param user - The user to upsert.
-   * @returns The upserted user.
-   */
-  async upsert(
-    user: {
-      /**
-       * Id of the user to insert.
-       */
-      id: string;
-      /**
-       * Email of the user to insert or new email to update for existing user.
-       */
-      email?: string | null | undefined;
-      /**
-       * Email verfified flag for the user to insert or new value to update for existing user.
-       */
-      emailVerified: boolean;
-      /**
-       * Accounts to insert for the user.
-       * Will be used only when the account is created. For existing users, the accounts will be ignored.
-       */
-      accounts: AccountInput[];
-      /**
-       * The extended public key used for locking proofs and mint quotes.
-       */
-      cashuLockingXpub: string;
-      /**
-       * The public key used for encryption.
-       */
-      encryptionPublicKey: string;
-      /**
-       * The user's Spark identity public key.
-       */
-      sparkIdentityPublicKey: string;
-      /**
-       * Timestamp when user accepted terms of service in ISO 8601 format.
-       * Optional because new OAuth users using the login flow are created before accepting TOS.
-       */
-      termsAcceptedAt?: string;
-      /**
-       * Timestamp when user accepted gift-card-mint terms of service in ISO 8601 format.
-       * Optional because it's only required when adding a mint that requires terms acceptance.
-       */
-      giftCardMintTermsAcceptedAt?: string;
-    },
-    options?: Options,
-  ): Promise<{ user: User; accounts: Account[] }> {
-    const accountsToAdd = user.accounts.map((account) => ({
-      name: account.name,
-      type: account.type,
-      currency: account.currency,
-      is_default: account.isDefault ?? false,
-      purpose: account.purpose,
-      details: (() => {
-        if (account.type === 'cashu') {
-          return CashuAccountDetailsDbDataSchema.parse({
-            mint_url: normalizeMintUrl(account.mintUrl),
-            is_test_mint: account.isTestMint,
-            keyset_counters: {},
-          } satisfies z.input<typeof CashuAccountDetailsDbDataSchema>);
-        }
-
-        return SparkAccountDetailsDbDataSchema.parse({
-          network: account.network,
-        } satisfies z.input<typeof SparkAccountDetailsDbDataSchema>);
-      })(),
-    }));
-
-    const query = this.db.rpc('upsert_user_with_accounts', {
-      p_user_id: user.id,
-      p_email: user.email ?? null,
-      p_email_verified: user.emailVerified,
-      p_accounts: accountsToAdd,
-      p_cashu_locking_xpub: user.cashuLockingXpub,
-      p_encryption_public_key: user.encryptionPublicKey,
-      p_spark_identity_public_key: user.sparkIdentityPublicKey,
-      p_terms_accepted_at: user.termsAcceptedAt,
-      p_gift_card_mint_terms_accepted_at: user.giftCardMintTermsAcceptedAt,
-    });
-
-    if (options?.abortSignal) {
-      query.abortSignal(options.abortSignal);
-    }
-
-    const { data, error } = await query;
-
-    if (error) {
-      throw new Error('Failed to upsert user', { cause: error });
-    }
-
-    const { accounts, user: upsertedUser } = data;
-    return {
-      user: ReadUserRepository.toUser(upsertedUser),
-      accounts: await Promise.all(
-        accounts.map((account) => this.accountRepository.toAccount(account)),
-      ),
-    };
-  }
-}
-
-export class ReadUserDefaultAccountRepository {
-  constructor(
-    private readonly db: AgicashDb,
-    private readonly queryClient: QueryClient,
-    private readonly getSparkWalletMnemonic: () => Promise<string>,
-    private readonly sparkStorageDir: string,
-  ) {}
-
-  /**
-   * Gets the user's default account. If currency is not provided, the default currency of the user is used.
-   * @returns The user's default account.
-   */
-  async getDefaultAccount(userId: string, currency?: Currency) {
-    const { data, error } = await this.db
-      .from('users')
-      .select(`
-        *,
-        accounts:accounts!user_id(
-          *,
-          cashu_proofs(*)
-        )
-      `)
-      .eq('id', userId)
-      .eq('accounts.cashu_proofs.state', 'UNSPENT')
-      .single();
-
-    if (error) {
-      throw new Error('Failed to get user default account IDs', error);
-    }
-
-    const defaultBtcAccountId = data.default_btc_account_id;
-    const defaultUsdAccountId = data.default_usd_account_id;
-
-    const accountCurrency = currency ?? data.default_currency;
-
-    const defaultAccountId =
-      accountCurrency === 'BTC' ? defaultBtcAccountId : defaultUsdAccountId;
-
-    const account = data.accounts.find(
-      (account) => account.id === defaultAccountId,
-    );
-
-    if (!account) {
-      throw new Error('No default account found for user');
-    }
-
-    return await this.toAccount(account);
-  }
-
-  private async toAccount(data: AgicashDbAccount): Promise<RedactedAccount> {
-    const commonData = {
-      id: data.id,
-      name: data.name,
-      currency: data.currency,
-      purpose: data.purpose,
-      state: data.state,
-      createdAt: data.created_at,
-      version: data.version,
-      expiresAt: data.expires_at,
-    };
-
-    if (isCashuAccount(data)) {
-      const details = data.details;
-
-      const { wallet, isOnline } = await getInitializedCashuWallet({
-        queryClient: this.queryClient,
-        mintUrl: details.mint_url,
-        currency: data.currency,
-        authProvider: getMintAuthProvider(data.purpose),
-      });
-
-      return {
-        ...commonData,
-        isOnline,
-        type: 'cashu',
-        mintUrl: details.mint_url,
-        isTestMint: details.is_test_mint,
-        keysetCounters: details.keyset_counters,
-        wallet,
-      };
-    }
-
-    if (isSparkAccount(data)) {
-      const { network } = data.details;
-      const { wallet, balance, isOnline } =
-        await this.getInitializedSparkWallet(network);
-      return {
-        ...commonData,
-        type: 'spark',
-        balance,
-        network,
-        isOnline,
-        wallet,
-      };
-    }
-
-    throw new Error('Invalid account type');
-  }
-
-  private async getInitializedSparkWallet(network: SparkNetwork) {
-    const mnemonic = await this.getSparkWalletMnemonic();
-    return getInitializedSparkWallet(
-      this.queryClient,
-      mnemonic,
-      network,
-      this.sparkStorageDir,
-    );
-  }
-}
-
-export class ReadUserRepository {
-  constructor(private readonly db: AgicashDb) {}
-
-  /**
-   * Gets a user from the database.
-   * @param userId - The id of the user to get.
-   * @returns The user.
-   */
-  async get(
-    userId: string,
-    options?: { abortSignal?: AbortSignal },
-  ): Promise<User> {
-    const query = this.db.from('users').select().eq('id', userId);
-
-    if (options?.abortSignal) {
-      query.abortSignal(options.abortSignal);
-    }
-
-    const { data, error } = await query.single();
-
-    if (error) {
-      throw new Error('Failed to get user', { cause: error });
-    }
-
-    return ReadUserRepository.toUser(data);
-  }
-
-  async getByUsername(
-    username: string,
-    options?: { abortSignal?: AbortSignal },
-  ): Promise<User | null> {
-    const query = this.db.from('users').select().eq('username', username);
-
-    if (options?.abortSignal) {
-      query.abortSignal(options.abortSignal);
-    }
-
-    const { data, error } = await query.maybeSingle();
-
-    if (error) {
-      throw new Error('Failed to get user by username', { cause: error });
-    }
-
-    return data ? ReadUserRepository.toUser(data) : null;
-  }
-
-  static toUser(dbUser: AgicashDbUser): User {
-    const commonData = {
-      id: dbUser.id,
-      username: dbUser.username,
-      emailVerified: dbUser.email_verified,
-      createdAt: dbUser.created_at,
-      updatedAt: dbUser.updated_at,
-      cashuLockingXpub: dbUser.cashu_locking_xpub,
-      encryptionPublicKey: dbUser.encryption_public_key,
-      sparkIdentityPublicKey: dbUser.spark_identity_public_key,
-      defaultBtcAccountId: dbUser.default_btc_account_id ?? '',
-      defaultUsdAccountId: dbUser.default_usd_account_id,
-      defaultCurrency: dbUser.default_currency,
-      termsAcceptedAt: dbUser.terms_accepted_at,
-      giftCardMintTermsAcceptedAt: dbUser.gift_card_mint_terms_accepted_at,
-    };
-
-    if (dbUser.email) {
-      return { ...commonData, email: dbUser.email, isGuest: false };
-    }
-
-    return { ...commonData, isGuest: true };
-  }
-}
 
 export function useReadUserRepository() {
   return new ReadUserRepository(agicashDbClient);
